可装，运行图片绘制不会出来

+++ src/com/niejun/androidgame/framework/Graphics.java
@@ -23,7 +23,7 @@ public interface Graphics {
 
 	public void drawRect(int x, int y, int width, int height, int color);
 
-	public void drawPixmap(Pixmap pixmap, float x, int y, int srcX, int srcY,
+	public void drawPixmap(Pixmap pixmap, int x, int y, int srcX, int srcY,
 			int srcWidth, int srcHeight);
 
 	public void drawPixmap(Pixmap pixmap, int x, int y);
